refactor(functions): move functions to a separated folder
